% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.tme.R
\name{dist.tme}
\alias{dist.tme}
\title{Distance function to components of the tumour microenvironment}
\usage{
dist.tme()
}
\arguments{
\item{hf}{A hyperframe}
}
\value{
Adds distance functions to the hyperframe
}
\description{
Distance function to components of the tumour microenvironment
}
\examples{
dist.tme()
}
