% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.gen.R
\name{marrow.gen}
\alias{marrow.gen}
\title{marrow.gen}
\usage{
marrow.gen()
}
\value{
A hyperframe
}
\description{
Generate a hyperframe of bone marrow representations from exported Qupath annotations and points.
}
\examples{
marrow.gen()
}
