% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.halfK.R
\name{marrow.halfK}
\alias{marrow.halfK}
\title{Test of homogeneity within each single point pattern}
\usage{
marrow.halfK(hf = hf, Kest = Kest)
}
\arguments{
\item{hf}{A hyperframe}

\item{Kest}{A summary function}
}
\value{
Plots of summary function with point patterns split into halves both ways.
}
\description{
Test of homogeneity within each single point pattern
}
\examples{
marrow.halfK(hf, Kest)
}
