% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.import.R
\name{marrow.import}
\alias{marrow.import}
\title{Import and clean}
\usage{
marrow.import()
}
\value{
A loaded hyperframe of data from marrow.gen
}
\description{
Import and clean
}
\examples{
marrow.import()
}
