% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.load.R
\name{marrow.load}
\alias{marrow.load}
\title{marrow.load}
\usage{
marrow.load()
}
\value{
Loaded hyperframe
}
\description{
Load a hyperframe created by the marrow.gen function
}
\examples{
marrow.load()
}
\author{
Callum Jeevan Fernando
}
