% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.rescale.R
\name{marrow.rescale}
\alias{marrow.rescale}
\title{Rescale hf}
\usage{
marrow.rescale(hf)
}
\arguments{
\item{hf}{A hyperframe}
}
\value{
A rescaled hyperframe of data
}
\description{
Rescale hf
}
\examples{
marrow.rescale(hf)
}
