% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.simpleplot.R
\name{marrow.simpleplot}
\alias{marrow.simpleplot}
\title{marrow.simpleplot}
\usage{
marrow.simpleplot()
}
\value{
a collection of plots
}
\description{
marrow.simpleplot
}
\examples{
marrow.simpleplot()
}
