% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrow.split.R
\name{marrow.split}
\alias{marrow.split}
\title{Split into test groups}
\usage{
marrow.split(hf)
}
\arguments{
\item{hf}{A hyperframe of point patterns with factor groups}
}
\value{
Multiple hyperframes split by group factor
}
\description{
Split into test groups
}
\examples{
marrow.split(hf)
}
