% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marrowgen.R
\name{marrowgen}
\alias{marrowgen}
\title{Generate a hyperframe of bone marrow representations from exported Qupath annotations and points.}
\usage{
marrowgen()
}
\value{
A hyperframe
}
\description{
Generate a hyperframe of bone marrow representations from exported Qupath annotations and points.
}
\examples{
marrowgen()
}
