% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{montecarlo}
\alias{montecarlo}
\title{Monte carlo test for difference in spatial covariates between test groups}
\usage{
montecarlo(
  points = ckit,
  group = group,
  spatcov = dist_adipo,
  hf = hf,
  geyr = 15,
  geysat = 3,
  nsim = 20
)
}
\arguments{
\item{points}{Points to be analysed (default is ckit)}

\item{group}{Factor group in hyperframe/name of column in hyperframe that determines sample group.}

\item{spatcov}{Spatial covariate}

\item{hf}{A hyperframe}

\item{geyr}{Geyer r variable}

\item{geysat}{Geyer saturation variable}

\item{nsim}{Number of simulations}
}
\value{
A monte carlo test result with p value
}
\description{
Monte carlo test for difference in spatial covariates between test groups
}
\examples{
montecarlo(points)
}
