% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.num.R
\name{row.num}
\alias{row.num}
\title{Appropriate row numbers}
\usage{
row.num(hf = hf)
}
\arguments{
\item{hf}{A hyperframe}
}
\value{
Renumber the row of hyperframe
}
\description{
Appropriate row numbers
}
\examples{
row.num(hf)
}
