% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.plot.R
\name{split.plot}
\alias{split.plot}
\title{Split plots of bone marrow point pattern hyperframe}
\usage{
\method{split}{plot}(spl)
}
\arguments{
\item{spl}{A list of split hyperframes generated from marrow.split()}
}
\value{
Split plots of hyperframe
}
\description{
Split plots of bone marrow point pattern hyperframe
}
\examples{
split.plot(spl)
}
