% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmark.ckit.R
\name{unmark.ckit}
\alias{unmark.ckit}
\title{Unmark the ckit column}
\usage{
\method{unmark}{ckit}(hf)
}
\arguments{
\item{hf}{A hyperframe}
}
\value{
An unmarked ckit column for hyperframe
}
\description{
Unmark the ckit column
}
\examples{
unmark.ckit(hf = hf)
}
